+++ src/main/resources/aether.mixins.json
@@ -13,8 +13,8 @@
     "accessors.SlotArmorAccessor",
     "accessory.ContainerInventoryMixinAccessory",
     "accessory.MenuInventoryMixinAddSlotAdjSlot",
-    "accessory.cape.invisibilitycape.MobPathfinderMixinForgetPlayerWhenToFar",
-    "accessory.cape.invisibilitycape.PlayerMixinInvisibility",
+    "accessory.cape.invisibilitycape.target.MobPathfinderMixinForgetPlayerWhenToFar",
+    "accessory.cape.invisibilitycape.player.PlayerMixinInvisibility",
     "accessory.gloves.ContainerInventoryMixinGlovesDamage",
     "accessory.quiver.ItemBowMixinQuiverSlotFix",
     "accessory.quiver.PlayerMixinNextArrow",
@@ -88,9 +88,9 @@
     "accessory.ScreenInventoryMixinNewInv",
     "accessory.cape.AetherDevCapeMixin",
     "accessory.cape.MobRendererPlayerMixinCapeRender",
-    "accessory.cape.invisibilitycape.EntityRendererMixinRemoveShadow",
-    "accessory.cape.invisibilitycape.MobRenderPlayerMixinInvisibility",
-    "accessory.cape.invisibilitycape.PlayerLocalMixinInvisibility",
+    "accessory.cape.invisibilitycape.render.EntityRendererMixinRemoveShadow",
+    "accessory.cape.invisibilitycape.render.MobRenderPlayerMixinInvisibility",
+    "accessory.cape.invisibilitycape.player.PlayerLocalMixinInvisibility",
     "accessory.gloves.MobRendererMixinExtendArmor",
     "accessory.quiver.MobRendererPlayerMixinCapeQuiver",
     "accessory.trinket.ItemElementMixinHoverShowSlot",
